broadcast_event: accept impl Into<u32> for the mask argument
The broadcast_event helper took a raw u32 and every call site spelled
out u32::from(EventMask::FOO). Change the signature to take
`impl Into<u32>` so callers can pass an EventMask directly (or still a
raw u32 — anything with `Into<u32>` works). The sed pass updates the
~25 call sites across property, window, color, drawing, and input
handlers; no behavior change.

+++ crates/sidecar/src/xserver/handlers/color.rs
@@ -153,7 +153,7 @@ pub(crate) fn handle_install_colormap(state: &mut ClientState, req: &InstallColo
         }, state.msb_first);
         state.pending_events.push(event.clone());
         // Also broadcast to other connections selecting on this window
-        state.broadcast_event(wid, u32::from(EventMask::COLOR_MAP_CHANGE), &event);
+        state.broadcast_event(wid, EventMask::COLOR_MAP_CHANGE, &event);
     }
     Vec::new()
 }
@@ -193,7 +193,7 @@ pub(crate) fn handle_uninstall_colormap(state: &mut ClientState, req: &Uninstall
             state: XColormapState::UNINSTALLED,
         }, state.msb_first);
         state.pending_events.push(event.clone());
-        state.broadcast_event(wid, u32::from(EventMask::COLOR_MAP_CHANGE), &event);
+        state.broadcast_event(wid, EventMask::COLOR_MAP_CHANGE, &event);
     }
 
     // Per spec, when a colormap is uninstalled, the default colormap
@@ -216,7 +216,7 @@ pub(crate) fn handle_uninstall_colormap(state: &mut ClientState, req: &Uninstall
             state: XColormapState::INSTALLED,
         }, state.msb_first);
         state.pending_events.push(event.clone());
-            state.broadcast_event(wid, u32::from(EventMask::COLOR_MAP_CHANGE), &event);
+            state.broadcast_event(wid, EventMask::COLOR_MAP_CHANGE, &event);
         }
     }
     Vec::new()
